Change the 2D terrain
Padded the terrain so that the drone starts in the center of the heightmap, but all of the elevated features are in front of it. Also changed the terrain to test valley with hills, a big climb, a cliff, and a taller hill (and inverse that going in the opposite direction).