Add full run demo and template mission
Full run shows how to run a mission while keeping track of your mission progress, adding waypoints to the mission in progress, checking lidar returns, registering the returns, and converting between world frame and lla. The template does all the same, but stubs out the part where the waypoints are calculated. Neither does terrain avoidance out of the box as that will give away the answer to the competition.